# docs/openapi.yaml
openapi: 3.0.0
servers:
  - description: CERCLES API
    url: http://localhost:8080
info:
  version: "1.0.0"
  title: CERCLES API Documentation
  description: >-
    Esta API proporciona los endpoints para la aplicación CERCLES.
components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Ingrese su token JWT para autenticar las solicitudes"
security:
  - BearerToken: []
...
paths:
...
  /api/evaluaciones/crear:
    post:
      tags:
        - Evaluación
      summary: Crear una nueva evaluación
      description: >
        Este endpoint permite a un profesor crear una nueva evaluación asociada a un curso. 
        Requiere proporcionar las fechas de inicio y fin de la evaluación, así como el identificador del curso correspondiente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fechaInicio:
                  type: string
                  format: date
                  description: Fecha de inicio de la evaluación en formato `YYYY-MM-DD`.
                fechaFin:
                  type: string
                  format: date
                  description: Fecha de fin de la evaluación en formato `YYYY-MM-DD`.
                cursoId:
                  type: integer
                  description: Identificador único del curso al que pertenece la evaluación.
              required:
                - fechaInicio
                - fechaFin
                - cursoId
      responses:
        "200":
          description: Evaluación creada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: Identificador único de la evaluación creada.
                  fechaInicio:
                    type: string
                    format: date
                    description: Fecha de inicio de la evaluación.
                  fechaFin:
                    type: string
                    format: date
                    description: Fecha de fin de la evaluación.
                  cursoId:
                    type: integer
                    description: Identificador único del curso asociado.
        "403":
          description: Usuario no autenticado
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no autenticado.
        "400":
          description: Error en los datos proporcionados
          content:
            text/plain:
              schema:
                type: string
                example: Formato incorrecto de las fechas o curso no encontrado.
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error al crear la evaluación.
  /api/evaluaciones/equipo/{equipoId}:
    get:
      tags:
        - Evaluación
      summary: Obtener evaluaciones realizadas por equipo
      description: >
        Este endpoint permite obtener las evaluaciones realizadas para un equipo específico. 
        Las evaluaciones pueden ser filtradas por una lista de identificadores opcionales.
      parameters:
        - name: equipoId
          in: path
          required: true
          description: Identificador único del equipo.
          schema:
            type: integer
        - name: evaluacionIds
          in: query
          required: false
          description: >
            Lista opcional de identificadores de evaluaciones para filtrar los resultados.
          schema:
            type: array
            items:
              type: integer
      responses:
        "200":
          description: Evaluaciones obtenidas exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    evaluadoId:
                      type: integer
                      description: Identificador único del estudiante evaluado.
                    evaluaciones:
                      type: array
                      items:
                        type: object
                        properties:
                          evaluacionId:
                            type: integer
                            description: Identificador único de la evaluación.
                          detallesEvaluacion:
                            type: array
                            items:
                              type: object
                              properties:
                                evaluadorId:
                                  type: integer
                                  description: Identificador único del estudiante evaluador.
                                puntos:
                                  type: integer
                                  description: Puntos otorgados en la evaluación.
        "403":
          description: Usuario no autenticado
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no autenticado.
        "400":
          description: Parámetros inválidos
          content:
            text/plain:
              schema:
                type: string
                example: Identificador del equipo no válido.
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error al obtener las evaluaciones.
  /api/evaluaciones/equipo/{equipoId}/medias:
    get:
      tags:
        - Evaluación
      summary: Obtener medias de evaluaciones por equipo
      description: >
        Este endpoint permite calcular y obtener las medias de las evaluaciones realizadas para un equipo específico, 
        incluyendo las medias por cada evaluación y las medias generales de los compañeros y autoevaluaciones.
      parameters:
        - name: equipoId
          in: path
          required: true
          description: Identificador único del equipo.
          schema:
            type: integer
      responses:
        "200":
          description: Medias obtenidas exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    estudianteId:
                      type: integer
                      description: Identificador único del estudiante evaluado.
                    mediasPorEvaluacion:
                      type: array
                      items:
                        type: object
                        properties:
                          evaluacionId:
                            type: integer
                            description: Identificador único de la evaluación.
                          mediaDeCompañeros:
                            type: number
                            format: float
                            description: Media de los puntos asignados por los compañeros.
                          notaPropia:
                            type: number
                            format: float
                            description: Puntos asignados en la autoevaluación.
                    mediaGeneralDeCompañeros:
                      type: number
                      format: float
                      description: Media general de los puntos asignados por los compañeros en todas las evaluaciones.
                    mediaGeneralPropia:
                      type: number
                      format: float
                      description: Media general de los puntos asignados en autoevaluaciones.
        "403":
          description: Usuario no autenticado
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no autenticado.
        "400":
          description: Identificador de equipo no válido
          content:
            text/plain:
              schema:
                type: string
                example: Equipo no encontrado.
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error al obtener las medias.
  /api/evaluaciones/equipo/{equipoId}/evaluacion-activa:
    get:
      tags:
        - Evaluación
      summary: Verificar si hay una evaluación activa
      description: >
        Este endpoint permite verificar si hay una evaluación activa para un equipo específico.
      parameters:
        - name: equipoId
          in: path
          required: true
          description: Identificador único del equipo.
          schema:
            type: integer
      responses:
        "200":
          description: Evaluación activa verificada
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    activa:
                      type: boolean
                      description: Booleano de si activa o no
                    fechaIni:
                      type: string
                      description: fecha inicio
                    fechaFin:
                      type: string
                      description: fecha fin
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error al verificar la evaluación activa.
  /api/evaluaciones/estudiante/crear:
    post:
      tags:
        - Evaluación
      summary: Crear detalles de una evaluación
      description: >
        Este endpoint permite a un estudiante crear detalles de una evaluación para los miembros de su equipo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                evaluacionId:
                  type: integer
                  description: Identificador único de la evaluación.
                evaluadorId:
                  type: integer
                  description: Identificador único del estudiante que realiza la evaluación.
                detalles:
                  type: array
                  items:
                    type: object
                    properties:
                      evaluadoId:
                        type: integer
                        description: Identificador único del estudiante evaluado.
                      puntos:
                        type: integer
                        description: Puntos asignados al estudiante evaluado.
      responses:
        "200":
          description: Evaluación guardada correctamente
          content:
            text/plain:
              schema:
                type: string
                example: Evaluación guardada correctamente.
        "403":
          description: Usuario no autenticado
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no autenticado.
        "400":
          description: Error en la solicitud
          content:
            text/plain:
              schema:
                type: string
                example: La suma de puntos asignados no coincide con el máximo permitido.
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error al guardar la evaluación.
...